Fix mob pathfinding!
Changes:
- Fixed a critical game-freezing bug where mobs could force chunks to load and continue moving indefinitely even when far from the player. The game now no longer forces chunk loading when checking for tiles in the world, instead returning None for unloaded areas. As a result, mobs in unloaded chunks will remain idle until they have loaded tiles nearby, preventing performance issues from excessive chunk generation.
- Save system and regions now use pickle protocol version 5 for better speed and compression ('MCPY' magic number version is now 2).
- Temporarily removed NBT (TBT) implementation as it was unnecessary (pickle will continue to be used instead).
- Optimized terrain generation.

+++ main.py
@@ -56,7 +56,7 @@ def main() -> None:
     nano_time: float = 1000000000.0 / GAME_TICKS
 
     timer = time() * 1000
-    delta: int = 0
+    delta: float = 0.00
 
     running: bool = True
     drawing: bool = False
@@ -116,6 +116,7 @@ def main() -> None:
         if ((time() * 1000) - timer) > 1000:
             if Game.debug:
                 print(f"> render time: {last_screen:.2f}ms")
+                #print(f"> FPS: {clock.get_fps():.2f}")
 
             timer += 1000
